+++ AGENTS.md
@@ -137,7 +137,7 @@ Open pull requests with **`dev`** as the base branch by default.
 
 ### State management (Jotai)
 
-- All shared atoms live in **`src/store/app.ts`** and end with the `Atom` suffix (e.g. `activeProfilerReportAtom`). Components and hooks consume atoms — they do not declare new ones inline. Add new atoms in the section comment block matching their feature area.
+- Shared atoms live under **`src/store/`**, end with the `Atom` suffix (e.g. `activeProfilerReportAtom`), and are discoverable via **`src/store/app.ts`**. Prefer declaring new atoms in `app.ts` in the section comment block matching their feature area. When co-locating atoms with mutators in another `store/` module (e.g. `store/fileTransferRegistry.ts` to avoid a circular import), **re-export them from `app.ts`**. Components and hooks consume atoms — they do not declare new ones inline.
 - Prefer **`useAtomValue`** for read-only consumers and **`useSetAtom`** for write-only consumers; use **`useAtom`** when a component both reads and writes the same atom. Don't subscribe via `useAtom` if you only need one half of the tuple.
 - Use **`atomWithStorage`** from `jotai/utils` for user-preference flags that need to survive reloads — never reach for `localStorage` directly.
