+++ src/components/operation-details/DeviceOperationsFullRender.tsx
@@ -36,8 +36,8 @@ import MemoryTag from '../MemoryTag';
 import { toReadableLayout, toReadableShape, toReadableType } from '../../functions/formatting';
 import { BufferTypeToStringBufferType, StringBufferType } from '../../model/BufferType';
 import {
-    DEVICE_OPERATION_ANALYSIS_RESULT,
     DEVICE_OPERATION_ANALYSIS_RESULT_LABEL,
+    DeviceOperationAnalysisResult,
 } from '../../definitions/DeviceOperationAnalysis';
 import { analyseDeviceOperation } from '../../functions/analyseDeviceOperation';
 
@@ -273,7 +273,7 @@ function useDeviceOperationsFullRenderModel(args: {
                     const opAnalysisResult = analyseDeviceOperation(node.operation);
                     const opAnalysisLabel = DEVICE_OPERATION_ANALYSIS_RESULT_LABEL[opAnalysisResult];
                     const labelClass = classNames('device-operation-label', {
-                        'failed-op-analysis': opAnalysisResult !== DEVICE_OPERATION_ANALYSIS_RESULT.OK,
+                        'failed-op-analysis': opAnalysisResult !== DeviceOperationAnalysisResult.OK,
                     });
                     const label = (
                         <h4 className={labelClass}>
@@ -282,7 +282,7 @@ function useDeviceOperationsFullRenderModel(args: {
                                     className='operation-icon'
                                     size={13}
                                     intent={
-                                        opAnalysisResult === DEVICE_OPERATION_ANALYSIS_RESULT.NOOP
+                                        opAnalysisResult === DeviceOperationAnalysisResult.NOOP
                                             ? Intent.WARNING
                                             : Intent.SUCCESS
                                     }
