+++ src/definitions/DeviceOperationAnalysis.ts
@@ -2,12 +2,12 @@
 //
 // SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
 
-export enum DEVICE_OPERATION_ANALYSIS_RESULT {
+export enum DeviceOperationAnalysisResult {
     OK,
     NOOP,
 }
 
 export const DEVICE_OPERATION_ANALYSIS_RESULT_LABEL = {
-    [DEVICE_OPERATION_ANALYSIS_RESULT.OK]: '',
-    [DEVICE_OPERATION_ANALYSIS_RESULT.NOOP]: 'Operation appears to be a no-op and may be removable',
+    [DeviceOperationAnalysisResult.OK]: '',
+    [DeviceOperationAnalysisResult.NOOP]: 'Operation appears to be a no-op and may be removable',
 };
