+++ src/functions/analyseDeviceOperation.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
 
-import { DEVICE_OPERATION_ANALYSIS_RESULT } from '../definitions/DeviceOperationAnalysis';
+import { DeviceOperationAnalysisResult } from '../definitions/DeviceOperationAnalysis';
 import { DeviceOperationNode, TensorNode } from '../model/APIData';
 
 const DEVICE_OPERATION_NAME = {
@@ -11,7 +11,7 @@ const DEVICE_OPERATION_NAME = {
 
 export const analyseDeviceOperation = (operation?: DeviceOperationNode) => {
     if (!operation) {
-        return DEVICE_OPERATION_ANALYSIS_RESULT.OK;
+        return DeviceOperationAnalysisResult.OK;
     }
     const { params, inputs, outputs } = operation;
     const { name } = params;
@@ -21,9 +21,9 @@ export const analyseDeviceOperation = (operation?: DeviceOperationNode) => {
         const outputTensor = outputs?.[0] as TensorNode;
         if (inputTensor && outputTensor && inputTensor.params.shape && outputTensor.params.shape) {
             if (inputTensor.params.shape === outputTensor.params.shape) {
-                return DEVICE_OPERATION_ANALYSIS_RESULT.NOOP;
+                return DeviceOperationAnalysisResult.NOOP;
             }
         }
     }
-    return DEVICE_OPERATION_ANALYSIS_RESULT.OK;
+    return DeviceOperationAnalysisResult.OK;
 };
